packages/policy/v1/requirements.txt
omegaconf>=0.0.0
tensorboardX
matplotlib
scikit-learn
opencv-python
h5py
open3d
tqdm
trimesh
torchdiffeq
pytorch-fid-wrapper
wandb
git+https://github.com/patrick-kidger/torchcubicspline.git
